moved the function for testing consistency of cuts to gadf module, moved also the corresponding unit test. included the (before missing) test data for this test. Since the test files are small they are added directly to the repository
